+++ game-middleearthshadowofwar/CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ## [2.2.0] - 2026-01-28
 
-- Added: Support for Middle-earth-Mod-Loader - installer and button to download. Note that this loader has some bugs and is not recommended.
+- Added: Support for Middle-earth-Mod-Loader - installer and button to download. Note that this loader has some bugs and is not recommended
+- Fixed: DLL Loader will now only auto-download if neither DLL Loader nor Middle-earth-Mod-Loader are installed (in Vortex or manually in the game folder)
 
 ## [2.1.0] - 2025-12-22